+++ packages/lib/jsx/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.77](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-jsx@0.0.76...@opendaw/lib-jsx@0.0.77) (2026-03-10)
+
+**Note:** Version bump only for package @opendaw/lib-jsx
+
 ## [0.0.76](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-jsx@0.0.75...@opendaw/lib-jsx@0.0.76) (2026-03-04)
 
 **Note:** Version bump only for package @opendaw/lib-jsx